Pipeline/requirements-dev.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
commonmark==0.9.1
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dacite==1.9.2
databricks-sdk==0.74.0
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
docker==7.1.0
exceptiongroup==1.3.1
executing==2.2.1
faicons==0.2.2
fastapi==0.124.2
fastjsonschema==2.21.2
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.61.0
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
great-tables==0.20.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
htmltools==0.6.0
httpcore==1.0.9
httpx==0.28.1
huey==2.5.5
idna==3.11
ImageHash==4.3.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
lazypredict==0.2.16
lightgbm==4.6.0
llvmlite==0.45.1
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.0
matplotlib-inline==0.2.1
minify_html==0.18.1
mistune==3.1.4
mlflow==3.7.0
mlflow-skinny==3.7.0
mlflow-tracing==3.7.0
multimethod==1.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
numba==0.62.1
numpy==2.2.6
nvidia-nccl-cu12==2.28.9
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.2
pexpect==4.9.0
phik==0.12.5
pillow==12.0.0
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
six==1.17.0
smmap==5.0.2
soupsieve==2.8
SQLAlchemy==2.0.45
sqlparse==0.5.4
stack-data==0.6.3
starlette==0.50.0
statsmodels==0.14.6
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.3.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
visions==0.8.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.4
widgetsnbextension==4.0.15
wordcloud==1.9.4
xgboost==3.1.2
ydata-profiling==4.18.0
zipp==3.23.0

# PDF Processing Pipeline Requirements
PyMuPDF>=1.23.0        # PDF parsing and structure analysis (fitz)
PyPDF2>=3.0.0          # PDF metadata and basic parsing
pdfminer.six>=20221105 # Text extraction
pytesseract>=0.3.10    # OCR for image-based PDFs
pdf2image>=1.16.3      # Convert PDF to images for OCR
pyod>=1.1.0
optuna>=4.0.0
